HT_16/requirements.txt
Pillow==10.1.0
Requests==2.31.0
selenium==4.16.0
reportlab==4.0.8